# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "XData Competitive & Clinical Intelligence Agent API",
  "description": "Evidence-grounded, continuously updated life-sciences competitive intelligence. Decision support only: facts are evidence-linked; interpretations are labelled AI hypotheses.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/ask": {
   "post": {
    "tags": [
     "ask"
    ],
    "summary": "Ask Question",
    "operationId": "ask_question_api_v1_ask_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AskIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Ask Question Api V1 Ask Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/ask/stream": {
   "post": {
    "tags": [
     "ask"
    ],
    "summary": "Ask Stream",
    "description": "Server-sent events: planned -> retrieved -> generated -> validated -> answer (NFR-PERF-001: AI may stream).",
    "operationId": "ask_stream_api_v1_ask_stream_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AskIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/ask/sessions": {
   "get": {
    "tags": [
     "ask"
    ],
    "summary": "Sessions",
    "operationId": "sessions_api_v1_ask_sessions_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "type": "array",
         "title": "Response Sessions Api V1 Ask Sessions Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/ask/sessions/{session_id}": {
   "get": {
    "tags": [
     "ask"
    ],
    "summary": "Session Detail",
    "operationId": "session_detail_api_v1_ask_sessions__session_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "session_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Session Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Session Detail Api V1 Ask Sessions  Session Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/ask/answers/{artifact_id}/export": {
   "get": {
    "tags": [
     "ask"
    ],
    "summary": "Export Answer",
    "operationId": "export_answer_api_v1_ask_answers__artifact_id__export_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "artifact_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Artifact Id"
      }
     },
     {
      "name": "format",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "md",
        "docx",
        "pdf",
        "pptx"
       ],
       "type": "string",
       "default": "docx",
       "title": "Format"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AskIn": {
    "properties": {
     "question": {
      "type": "string",
      "maxLength": 2000,
      "minLength": 3,
      "title": "Question"
     },
     "session_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session Id"
     },
     "landscape_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Landscape Id"
     }
    },
    "type": "object",
    "required": [
     "question"
    ],
    "title": "AskIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}